Redux flow completed
