--- src/Actions/index.js
@@ -1,6 +1,0 @@
-let addId=0;
-export const addTodo=(text)=>({
-    type:'ADD_TASK',
-    id: addId ++,
-    text
-})
