+++ src/store/NewMeasure/action.js
@@ -1,0 +1,6 @@
+let addId=0;
+export const createMeasure=(text)=>({
+    type:'CREATE_MEASURE',
+    id: addId ++,
+    text
+})
