+++ src/store/NewMeasure/reducer.js
@@ -1,13 +1,13 @@
 const todos = (state = [], action) => {
   switch (action.type) {
-    case 'ADD_TASK':
+    case 'CREATE_MEASURE':
     console.log(action);
       return [
         ...state,
         {
           id: action.id,
           text: action.text,
-          completed: false
+          
         }
       ]
