+++ src/store/rootReducer.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux'
-import todos from './todos'
+import todos from './NewMeasure/reducer'
 //import visibilityFilter from './visibilityFilter'
 
 const todoApp = combineReducers({
